Refactored disabled_due_to_debug method, moved to utils
